+++ src/diagnostics/diagnostics.ts
@@ -6,11 +6,12 @@
 
 import * as vscode from 'vscode';
 import { TEMPLATE_LANGUAGE, getOptions } from '../extension';
+import { Errors } from './common';
 import { analysePreamble } from './preambleCheck';
 import { analyseQrc } from './qrcCheck';
 import { analyseQbn } from './qbnCheck';
 import { tableCheck } from './tableCheck';
-import { Errors } from './common';
+import { ParameterTypes } from '../language/static/parameterTypes';
 import { Quest } from '../language/quest';
 
 let timer: NodeJS.Timer | null = null;
@@ -30,14 +31,14 @@ export function makeDiagnosticCollection(context: vscode.ExtensionContext): vsco
     if (vscode.window.activeTextEditor) {
         const document = vscode.window.activeTextEditor.document;
         if (document.languageId === TEMPLATE_LANGUAGE) {
-            diagnosticCollection.set(document.uri, doDiagnostics(document));
+            updateDiagnostics(diagnosticCollection, document);
         }
     }
 
     // Do diagnostics for opened files
     context.subscriptions.push(vscode.workspace.onDidOpenTextDocument((document) => {
         if (document.languageId === TEMPLATE_LANGUAGE) {
-            diagnosticCollection.set(document.uri, doDiagnostics(document));
+            updateDiagnostics(diagnosticCollection, document);
         }
     }));
 
@@ -52,7 +53,7 @@ export function makeDiagnosticCollection(context: vscode.ExtensionContext): vsco
                 }
 
                 timer = setTimeout(() => {
-                    diagnosticCollection.set(e.document.uri, doDiagnostics(e.document));
+                    updateDiagnostics(diagnosticCollection, e.document);
                     timer = null;
                 }, delay);
             }
@@ -63,7 +64,7 @@ export function makeDiagnosticCollection(context: vscode.ExtensionContext): vsco
         // Do diagnostics only on save
         context.subscriptions.push(vscode.workspace.onDidSaveTextDocument((document) => {
             if (document.languageId === TEMPLATE_LANGUAGE) {
-                diagnosticCollection.set(document.uri, doDiagnostics(document));
+                updateDiagnostics(diagnosticCollection, document);
             }
         }));
     }
@@ -74,21 +75,47 @@ export function makeDiagnosticCollection(context: vscode.ExtensionContext): vsco
 /**
  * Makes diagostics for the given document.
  */
-function doDiagnostics(document: vscode.TextDocument) {
+async function updateDiagnostics(diagnosticCollection: vscode.DiagnosticCollection, document: vscode.TextDocument): Promise<void> {
+    let diagnostics: vscode.Diagnostic[];
 
     if (Quest.isTable(document.uri)) {
-        
+
         // Analyse table
-        return Array.from(tableCheck(document));
+        diagnostics = Array.from(tableCheck(document));
     } else {
-        
+
         // Analyse quest
         const context = Quest.get(document);
-        return [
+        diagnostics = [
             ...analysePreamble(context),
             ...(context.qrc.found ? analyseQrc(context) : failedAnalysis(context, 'QRC')),
-            ...(context.qbn.found ? analyseQbn(context) : failedAnalysis(context, 'QBN'))
+            ...(context.qbn.found ? analyseQbn(context) : failedAnalysis(context, 'QBN')),
         ];
+        await analyseQuestReferences(context, diagnostics);
+    }
+
+    diagnosticCollection.set(document.uri, diagnostics);
+}
+
+/**
+ * Checks references to quests from task actions and ensures they point to quests available in the workspace.
+ * @param context The document to build diagnostics for.
+ * @param diagnostics The array where diagnostics are pushed.
+ */
+async function analyseQuestReferences(context: Quest, diagnostics: vscode.Diagnostic[]): Promise<void> {
+    let quests: Quest[] | undefined;
+    for (const action of context.qbn.iterateActions()) {
+        const index = action.signature.findIndex(x => x.type === ParameterTypes.questName || x.type === ParameterTypes.questID);
+        if (index !== -1) {
+            if (quests === undefined) {
+                quests = await Quest.getAll();
+            }
+
+            const name = Quest.indexToName(action.signature[index].value);
+            if (!quests.find(x => x.getName() === name)) {
+                diagnostics.push(Errors.undefinedQuest(action.getRange(index), name));
+            }
+        }
     }
 }